(MIKE) wip: reorganized monomial order engine code
